修复注册后不自动登录的问题。修复一些ui上的问题。

+++ Tumcca/Tumcca/src/com/pineapple/mobilecraft/tumcca/manager/UserManager.java
@@ -121,7 +121,7 @@ public class UserManager {
 	 * @param userId
 	 * @return {@link Profile}
 	 */
-	public Profile getUserProfile(int userId)
+	public Profile getUserProfile(long userId)
 	{
 		String key = String.valueOf(userId);
 		if(mProfilesCache.hasKey(key)){
